--- ml_algorithms/dummy_code.py
@@ -1,23 +1,0 @@
-"""
-.. module:: dummy_module
-   :synopsis: Donec id malesuada justo, eu condimentum elit.
-.. moduleauthor:: Aenean lobortis <github.com/lobortis>
-"""
-
-
-def dummy_func(arg1, arg2):
-
-    """Summary line.
-
-    Extended description of function.
-
-    :param int arg1: Description of arg1.
-    :param str arg2: Description of arg2.
-
-    :raise ValueError: if arg1 is equal to arg2
-
-    :return: Description of return value
-
-    :rtype: bool
-    """
-    return 1 / (arg1 - arg2) > 1
