+++ setup.cfg
@@ -1,11 +1,11 @@
 [metadata]
-name = ml_algorithms 
-summary = Machine Learning algorithms implemented in Python
+name = touvlo 
+summary = ML algorithms implemented from scratch provided block by block
 description-file = README.md
 author = Benardi Nunes
 author-email = benardinunes@gmail.com
 license = MIT
-home-page = https://github.com/Benardi/ml_algorithms 
+home-page = https://github.com/Benardi/touvlo 
 requires-python = >=3.5
 classifier =
     Development Status :: 1 - Alpha
@@ -20,7 +20,7 @@ classifier =
     Topic :: Software Development :: Libraries :: Python Modules
 
 [files]
-packages = ml_algorithms
+packages = touvlo
     
 [build_sphinx]
 source-dir = docs/source
