+++ tests/test_lin_rg.py
@@ -4,10 +4,9 @@ import pytest
 from numpy.testing import assert_allclose
 from numpy import ones, zeros, float64, array, append, genfromtxt
 
-from ml_algorithms.lin_rg import (normal_eqn, cost_func,
-                                  reg_cost_func, grad,
-                                  reg_grad, predict, h)
-from ml_algorithms.utils import numerical_grad
+from touvlo.lin_rg import (normal_eqn, cost_func, reg_cost_func, grad,
+                           reg_grad, predict, h)
+from touvlo.utils import numerical_grad
 
 TESTDATA1 = os.path.join(os.path.dirname(__file__), 'data1.csv')
 TESTDATA2 = os.path.join(os.path.dirname(__file__), 'data2.csv')
