+++ touvlo/kmeans.py
