+++ tox.ini
@@ -25,7 +25,7 @@ commands =
 ignore = E722, W503
 builtins =_
 exclude =venv,docs,.git,.tox,dist,doc,*lib/python*,*egg,build
-application_import_names = ml_algorithms
+application_import_names = touvlo
 import-order-style = pep8
 
 [testenv:pep8]
